// src/types/windowData.ts
export type WindowData = {
    aspectRatio?: number
    topMargin?: number
    bottomMargin?: number
    leftMargin?: number
    rightMargin?: number
    showXAxis?:boolean
    showYAxis?:boolean
    showXGrid?:boolean
    showYGrid?:boolean
    color?: string
    fontSize?:number
    legend? :boolean
    title?:string
    xAxisLabel?: string
    yAxisLabel?: string
}